\docType{data}
\name{example}
\alias{rawcount}
\alias{rawcoord}

\title{Example Data of Human Breast Cancer}
\usage{
data(example)
}

\description{
\describe{
\item{rawcount}{A numeric matrix of gene expression counts for 14,798 genes and 251 spots.} \item{rawcoord}{A numeric matrix of two-dimensioanl spatial coordinates for 251 spots.}
}
}







