\name{fit_glmm}
\alias{fit_glmm}
\title{Internal Function of mixSVG}
\usage{
fit_glmm(y, X, model_init, libsize, n_iter = 1000, tol = 1e-4)
}
\description{
Internal function for estimation under the null.
}

\arguments{
\item{y}{}
\item{X}{}
\item{model_init}{}
\item{libsize}{}
\item{n_iter}{}
\item{tol}{}
}

\keyword{internal}










