\name{mixSVG_main}
\alias{mixSVG_main}
\title{Internal Function of mixSVG}
\usage{
mixSVG_main(y, X, s_trans, pat_idx, perm_sample, libsize, vtest_zero_prop)
}

\description{
Internal function for implementing the mixSVG method.
}

\arguments{
\item{y}{}
\item{X}{}
\item{s_trans}{}
\item{pat_idx}{}
\item{perm_sample}{}
\item{libsize}{}
\item{vtest_zero_prop}{}
}

\value{
\item{model0}{}
\item{pval}{}
\item{pval_pat}{}
}

\keyword{internal}










