\name{transcoord_func}
\alias{transcoord_func}
\title{Internal Function of mixSVG}
\usage{
transcoord_func(si, transfunc = "gaussian", q = 0.2, c = 0)
}
\description{
Internal function for performing transformation of spatial coordinates.
}

\arguments{
\item{si}{}
\item{transfunc}{}
\item{q}{}
\item{c}{}
}

\keyword{internal}










